move next question endpoint to FHIR controller
