+++ .vscode/launch.json
@@ -6,10 +6,23 @@
     "configurations": [
         {
             "type": "java",
-            "name": "Debug CRD (Local + Docker)", // gradle bootRun -Pdebug,
+            "name": "Launch Application",
+            "request": "launch",
+            "mainClass": "org.hl7.davinci.endpoint.Application",
+            "projectName": "server"
+        },
+        {
+            "type": "java",
+            "name": "Launch Java Program",
+            "request": "launch",
+            "mainClass": ""
+        },
+        {
+            "type": "java",
+            "name": "Debug CRD (Local + Docker)",
             "request": "attach",
             "hostName": "localhost",
-            "port": 8091,
+            "port": 8091
         }
     ]
 }
