+++ server/src/main/java/org/hl7/davinci/endpoint/fhir/r4/QuestionnaireNextQuestionOperation.java
@@ -1,5 +1,6 @@
-package org.hl7.davinci.endpoint.controllers;
+package org.hl7.davinci.endpoint.fhir.r4;
 
+import org.hl7.davinci.endpoint.files.FileStore;
 import org.hl7.davinci.endpoint.Application;
 import org.hl7.davinci.endpoint.files.FileResource;
 import org.hl7.davinci.endpoint.files.FileStore;
@@ -10,6 +11,7 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
 
 import ca.uhn.fhir.context.FhirContext;
@@ -42,25 +44,20 @@ import org.hl7.fhir.r4.model.QuestionnaireResponse.QuestionnaireResponseStatus;
 import org.hl7.fhir.r4.model.Reference;
 import org.hl7.davinci.endpoint.files.QuestionnaireEmbeddedCQLProcessor;
 
-// --- ORDER OF RESPONSE-REQUEST OPERATIONS
-// (REQUEST) External user sends the initial QuestionnaireResponse JSON that contains which questionnaire it would like to trigger as n element the "contained" field.
-// (RESPONSE) QuestionnaireController adds the first question with its answerResponse options (with its linkId and text) to the JSON in QuestionnaireResponse.contained.item[] and sends it back.
-// (REQUEST) External user adds their answer to the question to the JSON in QuestionnaireResponse.item[] and sends it back.
-// (RESPONSE) QuestionnaireController takes that response and adds the next indicated question to the JSON in QuestionnaireResponse.contained.item[] and sends it back.
-// Repeat intil QuestionnaireController reaches a leaf-node, then it sets the status to "completed" from "in-progress"
-// Ultimately, The QuestionnaireController responses add ONLY to the QuestionnaireResponse.contained.item[]. The external requester adds answers to QuestionnaireResponse.item[] and includes the associated linkid and text.
 
-@CrossOrigin
-@RestController
-@RequestMapping("/Questionnaire")
-public class QuestionnaireController {
-
-    @Autowired
-    private FileStore fileStore;
+public class QuestionnaireNextQuestionOperation {
+    
+    FileStore fileStore;
 
     private QuestionnaireEmbeddedCQLProcessor questionnaireEmbeddedCQLProcessor;
 
-    public QuestionnaireController() {
+     // Logger.
+     private static Logger logger = Logger.getLogger(Application.class.getName());
+     // Trees that track the current and next questions. Is key-value mappng of: Map<Questionnaire ID -> AdaptiveQuestionnaireTree>
+     private static final Map<String, AdaptiveQuestionnaireTree> questionnaireTrees = new HashMap<String, AdaptiveQuestionnaireTree>();
+
+    public QuestionnaireNextQuestionOperation(FileStore fileStore) {
+        this.fileStore = fileStore;
         this.questionnaireEmbeddedCQLProcessor = new QuestionnaireEmbeddedCQLProcessor();
     }
 
@@ -267,29 +264,14 @@ public class QuestionnaireController {
         }
     }
 
-    // Logger.
-    private static Logger logger = Logger.getLogger(Application.class.getName());
-    // Trees that track the current and next questions. Is key-value mappng of: Map<Questionnaire ID -> AdaptiveQuestionnaireTree>
-    private static final Map<String, AdaptiveQuestionnaireTree> questionnaireTrees = new HashMap<String, AdaptiveQuestionnaireTree>();
-
-    /**
-     * Retrieves the next question based on the request.
-     * @param request
-     * @param entity
-     * @return
-     */
-    @PostMapping(value = "/$next-question", consumes = { MediaType.APPLICATION_JSON_VALUE, "application/fhir+json" })
-    public ResponseEntity<String> retrieveNextQuestion(HttpServletRequest request, HttpEntity<String> entity) {
-        return getNextQuestionOperation(entity.getBody(), request);
-    }
-
-    /**
+    
+     /**
      * Returns the next question based on the request.
      * @param body
      * @param request
      * @return
      */
-    private ResponseEntity<String> getNextQuestionOperation(String body, HttpServletRequest request) {
+    public ResponseEntity<String> getNextQuestionOperation(String body, HttpServletRequest request) {
         logger.info("POST /Questionnaire/$next-question fhir+");
 
         FhirContext ctx = new FhirComponents().getFhirContext();
@@ -348,7 +330,7 @@ public class QuestionnaireController {
                 // Pull the resulting next question that the recieved responses and answers point to from the tree without including its children.
                 List<QuestionnaireItemComponent> nextQuestionSetResults = currentTree.getNextQuestionsForAnswers(allResponses, inputQuestionnaireResponse);
                 // Add the next set of questions to the response.
-                QuestionnaireController.addQuestionSetToQuestionnaireResponse(inputQuestionnaireResponse, nextQuestionSetResults);
+                QuestionnaireNextQuestionOperation.addQuestionSetToQuestionnaireResponse(inputQuestionnaireResponse, nextQuestionSetResults);
                 // Check that there no duplicates in the set of questions.
                 if ((new HashSet(((Questionnaire) inputQuestionnaireResponse.getContained().get(0)).getItem().stream().map(item -> item.getLinkId()).collect(Collectors.toList()))).size() != ((Questionnaire) inputQuestionnaireResponse.getContained().get(0)).getItem().size()){
                     throw new RuntimeException("Attempted to send a set of questions with duplicates. Question IDs are: " + (((Questionnaire) inputQuestionnaireResponse.getContained().get(0)).getItem().stream().map(item -> item.getLinkId()).collect(Collectors.toList())));
